+++ packages/xstate-graph/src/index.ts
@@ -9,11 +9,7 @@ export {
 export { getSimplePaths } from './simplePaths.ts';
 export { getShortestPaths } from './shortestPaths.ts';
 export { getPathsFromEvents } from './pathFromEvents.ts';
-export { getAdjacencyMap } from './adjacency.ts';
-export {
-  TestModel,
-  createTestModel,
-  adjacencyMapToArray
-} from './TestModel.ts';
+export { getAdjacencyMap, adjacencyMapToArray } from './adjacency.ts';
+export { TestModel, createTestModel } from './TestModel.ts';
 export * from './pathGenerators.ts';
 export * from './types.ts';