+++ packages/xstate-test/src/index.ts
@@ -1,3 +1,0 @@
-export { TestModel, createTestModel } from './TestModel.ts';
-export * from './types.ts';
-export * from './pathGenerators.ts';